incomplete unfortunately, but i'll have the complete one later, for what it's worth

+++ P1StoreApplication/P1Repository/P1MapperClass.cs
@@ -77,7 +77,7 @@ namespace P1Repository
                 OrderID = (int)reader5[0],
                 fk_StoreID = (int)reader5[1],
                 fk_CustomerID = (int)reader5[2],
-                total_cost = (int)reader5[3]
+                total_cost = (decimal)reader5[3]
             };
             return oh;
         }
